Added v0.01 to main repository
This is the first version available to download. This version showcases the creation of the title screen and the save loading function which is currently incomplete as I do not have a save object to program a proper save function to. I have created animations for the player's forward movement, and their forward idle! That's all for now!